Took nara1 branch fixed errors, got rid of unused files and code, updated readme, working minus company info and resizing chart bug
