+++ models/index.js
@@ -1,6 +1,5 @@
 const User = require('./User');
 const Portfolio = require('./Portfolio');
-const Ratio = require('./Ratio');
 
 User.hasMany(Portfolio, {
   foreignKey: 'user_id',
@@ -11,4 +10,4 @@ Portfolio.belongsTo(User, {
   foreignKey: 'user_id'
 });
 
-module.exports = { User, Portfolio, Ratio };
+module.exports = { User, Portfolio };
