--- public/js/index.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Ratio Test</title>
-</head>
-
-<body>
-    <h1>Testing Ratio data from financialmodelingprep.com</h1>
-
-    <div>
-
-
-        <div class="search-container">
-  
-
-                  <input class="searchbar" id="user-input" type="text" placeholder="Search.." name="search">
-
-                  <button type="submit" id="search-button"><i class="fa fa-search"></i></button>
-
-
-  </form>
-
-</div>
-
-<div>
-
-  <select class="form-control browser-default dropdown" id="ratioChoice" onchange="changeSelect()">
-            {{!-- <option value="" selected="selected">Choose Ratio</option> --}}
-            <option value="1">Current Ratio</option>
-            <option value="2">Working Capital</option>
-            <option value="3">Quick Ratio</option>
-          </select>
-
-</div>  
-
-    
-  <div>
-
-    {{!-- chart --}}
-      <canvas id="myChart" width="800" height="400"></canvas>
-
-  </div>
-
-  <div id="ratioText">The current ratio is a liquidity ratio that measures a companys ability to pay short-term obligations or those due within one year. It tells investors and analysts how a company can maximize the current assets on its balance sheet to satisfy its current debt and other payables. A current ratio that is in line with the industry average or slightly higher is generally considered acceptable. A current ratio that is lower than the industry average may indicate a higher risk of distress or default. Similarly, if a company has a very high current ratio compared with its peer group, it indicates that management may not be using its assets efficiently.</div>
-
-</div>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js"></script>
-<script src="jquery-3.6.0.min.js"></script>
-
-    <script src="./script.js"></script>
-</body>
-
-</html>
