+++ views/layouts/main.handlebars
@@ -1,52 +1,59 @@
 <!DOCTYPE html>
 <html lang="en">
 
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Financial Ratio Analysis</title>
-  <link rel="stylesheet" href="/css/jass.css">
-  <link rel="stylesheet" href="/css/style.css">
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com">
-  <link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&display=swap" rel="stylesheet">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-</head>
+  <head>
+    <meta charset='UTF-8' />
+    <meta name='viewport' content='width=device-width, initial-scale=1.0' />
+    <title>Financial Ratio Analysis</title>
+    <link rel='stylesheet' href='/css/jass.css' />
+    <link rel='stylesheet' href='/css/style.css' />
+    <link rel='preconnect' href='https://fonts.googleapis.com' />
+    <link rel='preconnect' href='https://fonts.gstatic.com' />
+    <link
+      href='https://fonts.googleapis.com/css2?family=Bebas+Neue&display=swap'
+      rel='stylesheet'
+    />
+    <link
+      rel='stylesheet'
+      href='https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css'
+    />
+  </head>
 
-<body>
-    <header class="display-flex justify-space-between align-center p-2">
+  <body>
+    <header class='display-flex justify-space-between align-center p-2'>
       <h1>
-        <a href="/">Financial Ratio Analysis </a>
+        <a href='/'>Financial Ratio Analysis </a>
       </h1>
-        <div class="navbar">
-          {{#if logged_in}}
-              <button class='no-button' id='logout'>logout</button>
-             {{else}}
-            <a href="/login">Login</a>
-             {{/if}}
+      <div class='navbar'>
+        {{#if logged_in}}
+          <button class='no-button' id='logout'>logout</button>
+        {{else}}
+          <a href='/login'>Login</a>
+        {{/if}}
 
-          </div>
+      </div>
     </header>
 
-    <main class="container container-fluid">
+    <main class='container container-fluid'>
 
-        {{{ body }}}
-        
+      {{{body}}}
+
     </main>
 
     <footer>
 
-      
     </footer>
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+    <script src='/js/search.js'></script>
+  
+    <!-- Render script for logged in users only -->
 
-  <!-- Render script for logged in users only -->
+    {{#if logged_in}}
 
-  {{#if logged_in}}
+      <script src='/js/logout.js'></script>
 
-  <script src="/js/logout.js"></script>
+    {{/if}}
 
-  {{/if}}
-
-</body>
+  </body>
 
 </html>
