+++ views/ratio.handlebars
@@ -1,46 +1,98 @@
+<div class="row">
+  <div class="col-auto">
+    <h2 id="user-info" data-id="{{user.id}}">Welcome, {{user.name}}!</h2>
+  </div>
+</div>
+<!-- Enter Company Name -->
+<div class="row">
+  <div class="col-auto">
+    <form>
+      <div class="form-group">
+        <label for="ticker">Portfolio:</label>
+        <input class="form-control" type="text" id="user-input" name="ticker" required value="AAPL">
+      </div>
+      <button id="add-button" class="form-control btn btn-search">Add</button>
+      <hr class="my-4">
+      <!-- Show Portfolio List -->
+      {{#if user.portfolios.length}}
+      <div class="form-group">
+        {{#each user.portfolios as |portfolio| }}
+        <label for="portfolio-checkbox">
+          <input type="checkbox" class="single-check" data-id="{{portfolio.id}}" name="portfolio-checkbox"
+            onclick="onlyOne(this)" /> {{portfolio.company_symbol}}
+        </label>
+        {{/each}}
+      </div>
+      <button id="delete-button" class="form-control btn btn-danger">Delete</button>
+      {{/if}}
+    </form>
+  </div>
+</div>
 
 
 <div>
 
+  <div>
 
-          <div class="search-container">
-    
-
-                    <input class="searchbar" id="user-input" type="text" placeholder="Search.." name="search">
-
-                    <button type="submit" id="search-button"><i class="fa fa-search"></i></button>
-
-
-    </form>
+    <select class="form-control browser-default dropdown" id="ratio-choice" onchange="changeSelect()">
+      {{!-- <option value="" selected="selected">Choose Ratio</option> --}}
+      <option value="1">Current Ratio</option>
+      <option value="2">Quick Ratio</option>
+      <option value="3">Working Capital</option>
+    </select>
 
   </div>
 
+{{!-- {{#if user.portfolios.length}} --}}
   <div>
 
-    <select class="form-control browser-default dropdown" id="ratioChoice" onchange="changeSelect()">
-              {{!-- <option value="" selected="selected">Choose Ratio</option> --}}
-              <option value="1">Current Ratio</option>
-              <option value="2">Working Capital</option>
-              <option value="3">Quick Ratio</option>
-            </select>
+    {{!-- chart --}}
+    {{!-- <div class="chart-container" style="position: relative; height:40vh; width:80vw"> --}}
+      {{!-- <canvas id="myChart"</canvas> --}}
+      <canvas id="myChart" width="800" height="400"></canvas>
+      {{!-- <p>Current Ratio</p> --}}
 
-  </div>  
+  </div>
 
-      
-    <div>
-
-      {{!-- chart --}}
-        <canvas id="myChart" width="800" height="400"></canvas>
-
-    </div>
-
-    <div id="ratioText">The current ratio is a liquidity ratio that measures a companys ability to pay short-term obligations or those due within one year. It tells investors and analysts how a company can maximize the current assets on its balance sheet to satisfy its current debt and other payables. A current ratio that is in line with the industry average or slightly higher is generally considered acceptable. A current ratio that is lower than the industry average may indicate a higher risk of distress or default. Similarly, if a company has a very high current ratio compared with its peer group, it indicates that management may not be using its assets efficiently.</div>
+  <div id="ratio-text">The current ratio is a liquidity ratio that measures a companys ability to pay short-term
+    obligations or those due within one year. It tells investors and analysts how a company can maximize the current
+    assets on its balance sheet to satisfy its current debt and other payables. A current ratio that is in line with the
+    industry average or slightly higher is generally considered acceptable. A current ratio that is lower than the
+    industry average may indicate a higher risk of distress or default. Similarly, if a company has a very high current
+    ratio compared with its peer group, it indicates that management may not be using its assets efficiently.</div>
 
 </div>
+{{!-- {{/if}} --}}
 
-      
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js"></script>
-    <script src="/js/ratio.js"></script>
-    <script src="/js/ratiopage.js"></script>
-    <script src="jquery-3.6.0.min.js"></script>
-    
+
+{{#if user.portfolios.length}}
+<div class="row">
+  <div class="col-auto">
+    <p class="text-center"> Company Information </p>
+  </div>
+</div>
+
+<button id="company-button" class="btn btn-primary">Information</button>
+<button id="result-button" class="btn btn-primary">Show Result</button>
+
+<div class="row">
+  <div class="col-auto">
+    <div class="card">
+      {{#each user.portfolios as |portfolio| }}
+      <div class="card-header">
+        <h4>Company: companyName {{portfolio.company_symbol}}</h4>
+        <p></p>
+      </div>
+      <div class="card-body">
+        <p class="card-text"></p>
+      </div>
+      {{/each}}
+    </div>
+  </div>
+</div>
+{{/if}}
+
+
+{{!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js"></script> --}}
+<script src="//cdnjs.cloudflare.com/ajax/libs/Chart.js/3.7.1/chart.min.js"></script>
+<script src="/js/ratio.js"></script>
